Added adapter, decorator, test decorator

+++ src/main/java/sda/mvc/model/Model.java
@@ -4,13 +4,13 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Model {
-    public static List<Integer> values = new ArrayList<>();
+    public static List<Long> values = new ArrayList<>();
 
-    public static List<Integer> getValues() {
+    public static List<Long> getValues() {
         return values;
     }
 
-    public static void addValue(int newValue) {
+    public static void addValue(long newValue) {
         values.add(newValue);
     }
 }
